refactor(fathers-page): extract  into a separate component for Fathers and Figures page
